zkclient删除节点时首先判断该节点是否存在，并打印信息

+++ dubbo-remoting/dubbo-remoting-zookeeper/src/main/java/com/alibaba/dubbo/remoting/zookeeper/zkclient/ZkclientZookeeperClient.java
@@ -54,8 +54,13 @@ public class ZkclientZookeeperClient extends AbstractZookeeperClient<IZkChildLis
 
 	public void delete(String path) {
 		try {
-			client.delete(path);
+			 if(client.exists(path)){
+				 client.delete(path);
+			 }else{
+				 System.out.println("节点不存在:"+path);
+			 }
 		} catch (ZkNoNodeException e) {
+			System.out.println(e.getMessage());
 		}
 	}
